requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.0
transformers==4.35.0
spacy==3.7.2

# IBM watsonx.ai
ibm-watson-machine-learning==1.0.327

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0

# Jira API
requests==2.31.0

# Slack API (requests is used, but you can also use slack-sdk)
# slack-sdk==3.23.0

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

